feat: SSE optimization, dynamic paging, scheduler integration and desktop rendering
Major architectural improvements across memory management, process scheduling and GUI rendering.

MEMORY & PERFORMANCE
- Implement CPUID-based SSE/SSE2 detection with dynamic enablement
- Add memcpy_sse(): 128-bit XMM-optimized memory copy (2x throughput on SSE systems)
- Refactor physical memory manager: fix bitmap alignment and absolute addressing

PROCESS ISOLATION & PAGING
- Implement dynamic page directory allocation per process (CreateProcessDirectory)
- Introduce MapPage() primitive with support for Present, RW, User, Dirty flags
- Replace static 1024x1024 arrays with PMM-allocated dynamic structures
- Update memory layout: 0-256MB kernel, 256MB-3GB heap, 3GB-4GB MMIO

INTERRUPT & SCHEDULER INTEGRATION
- Refactor InterruptManager constructor: ProcessManager → Scheduler dependency
- Add InterruptGuard RAII class for atomic interrupt operations
- Integrate KernelSymbolTable::PrintStackTrace() for crash debugging
- Implement PIC EOI handling in DoHandleInterrupt()

DESKTOP RENDERING OPTIMIZATION
- Implement two-pass rendering: full redraw + mouse-only updates
- Add per-process event handler registration (Desktop::createNewHandler)
- Introduce 20×20 cursor back-buffer for flicker-free rendering
- Achieve ~80% reduction in full-screen desktop redraws

WIDGET SYSTEM REFACTORING
- Remove hardcoded 100-child limit for unlimited widget support
- Standardize focus API: setFocus() → SetFocus(), isFocussed → isFocused
- Enhance GetFocus() with PushBack() for proper Z-ordering
- Improve hit testing with ReverseForEach() for top-most-first detection

GUI COMPONENT IMPROVEMENTS
- Button: MarkDirty() optimization for cache management
- Label, InputBox, MessageBox: Enhanced rendering and text handling
- Window: Improved drag detection and coordinate conversion
- Cache initialization: Fixed in constructors for all components

AUDIO SUBSYSTEM
- Integrate AudioMixer with global reference (g_systemMixer)
- Support audio driver initialization (g_systemcAudioDriver)
- Implement 10ms audio tick counter for mixer updates
- Add WAV file support to filesystem and HDD provisioning

FILESYSTEM & DRIVERS
- Add Doxygen headers to FAT32.cpp, File.cpp, msdospart.cpp, ata.cpp
- Standardize pointer syntax across filesystem module
- Normalize spacing in conditional statements and declarations

BUILD & DEVELOPER EXPERIENCE
- Add frame pointer preservation for improved stack traces
- Implement linker map generation for symbol debugging
- Add code quality targets: check-style, check-bugs, check-eof, check-headers
- Update user program memory layout: 0x10001000 (256MB + guard page)

UTILITIES & USER PROGRAMS
- Add string utilities: atoi(), HexStrToInt(), strncmp()
- Restructure user_prog directory: extract common code to libhx86 static library
- Build libhx86.a separately with dependency chain enforcement
- Link MeMView and test program against shared libhx86 library
- Update user program memory layout: 0x10001000 (256MB + guard page)
- Update AudioMixer and GraphicsDriver global references
- Add PIT port definitions for timing support

+++ core/drivers/ata.cpp
@@ -1,77 +1,79 @@
+/**
+ * @file        ata.cpp
+ * @brief       ATA Driver Implementation
+ *
+ * @date        01/02/2026
+ * @version     1.0.0
+ */
+
 #include <core/drivers/ata.h>
 
-
 AdvancedTechnologyAttachment::AdvancedTechnologyAttachment(bool master, uint16_t portBase)
-:   dataPort(portBase),
-    errorPort(portBase + 0x1),
-    sectorCountPort(portBase + 0x2),
-    lbaLowPort(portBase + 0x3),
-    lbaMidPort(portBase + 0x4),
-    lbaHiPort(portBase + 0x5),
-    devicePort(portBase + 0x6),
-    commandPort(portBase + 0x7),
-    controlPort(portBase + 0x206)
-{
+    : dataPort(portBase),
+      errorPort(portBase + 0x1),
+      sectorCountPort(portBase + 0x2),
+      lbaLowPort(portBase + 0x3),
+      lbaMidPort(portBase + 0x4),
+      lbaHiPort(portBase + 0x5),
+      devicePort(portBase + 0x6),
+      commandPort(portBase + 0x7),
+      controlPort(portBase + 0x206) {
     this->master = master;
 }
 
-AdvancedTechnologyAttachment::~AdvancedTechnologyAttachment()
-{
-}
+AdvancedTechnologyAttachment::~AdvancedTechnologyAttachment() {}
 
 // Returns the total number of sectors on the drive (LBA28)
 // Returns 0 if error or drive not present
-uint32_t AdvancedTechnologyAttachment::Identify()
-{
-    printf("Identifying %s %s drive...\n", ((this->dataPort.getPortNumber() == 0x1F0) ? "primary" : "secondary"), (master ? "master" : "slave"));
+uint32_t AdvancedTechnologyAttachment::Identify() {
+    printf("Identifying %s %s drive...\n",
+           ((this->dataPort.getPortNumber() == 0x1F0) ? "primary" : "secondary"),
+           (master ? "master" : "slave"));
 
     devicePort.Write(master ? 0xA0 : 0xB0);
     controlPort.Write(0);
-    
+
     devicePort.Write(0xA0);
     uint8_t status = commandPort.Read();
-    if(status == 0xFF) return 0; // No Device
-    
+    if (status == 0xFF) return 0;  // No Device
+
     devicePort.Write(master ? 0xA0 : 0xB0);
     sectorCountPort.Write(0);
     lbaLowPort.Write(0);
     lbaMidPort.Write(0);
     lbaHiPort.Write(0);
-    commandPort.Write(0xEC); // Identify command
-    
+    commandPort.Write(0xEC);  // Identify command
+
     status = commandPort.Read();
-    if(status == 0x00) return 0; // No Device
-    
-    while(((status & 0x80) == 0x80) && ((status & 0x01) != 0x01))
-        status = commandPort.Read();
-        
-    if(status & 0x01) {
+    if (status == 0x00) return 0;  // No Device
+
+    while (((status & 0x80) == 0x80) && ((status & 0x01) != 0x01)) status = commandPort.Read();
+
+    if (status & 0x01) {
         printf("ATA IDENTIFY ERROR\n");
         return 0;
     }
-    
+
     uint32_t totalSectors = 0;
-    
-    for(int i = 0; i < 256; i++)
-    {
+
+    for (int i = 0; i < 256; i++) {
         uint16_t data = dataPort.Read();
-        
+
         // Words 60 and 61 contain the total sector count for LBA28
-        if(i == 60) {
+        if (i == 60) {
             totalSectors = data;
-        }
-        else if(i == 61) {
+        } else if (i == 61) {
             totalSectors |= ((uint32_t)data << 16);
         }
     }
-    
-    printf("HDD Identified. Size: %d Sectors (%d MB)\n", totalSectors, (totalSectors * 512) / 1024 / 1024);
+
+    printf("HDD Identified. Size: %d Sectors (%d MB)\n", (int32_t)totalSectors,
+           (int32_t)(totalSectors * 512) / 1024 / 1024);
     this->ata_size = totalSectors;
     return totalSectors;
 }
 
-void AdvancedTechnologyAttachment::Read28(uint32_t sectorNum, uint8_t* data, int count)
-{
+void AdvancedTechnologyAttachment::Read28(uint32_t sectorNum, uint8_t* data, int count) {
     if (sectorNum > 0x0FFFFFFF) return;
 
     devicePort.Write((master ? 0xE0 : 0xF0) | ((sectorNum & 0x0F000000) >> 24));
@@ -88,7 +90,10 @@ void AdvancedTechnologyAttachment::Read28(uint32_t sectorNum, uint8_t* data, int
     uint8_t status4 = commandPort.Read();
 
     while ((status & 0x80) == 0x80) status = commandPort.Read();
-    if ((status & 0x01) == 0x01) { printf("ATA READ ERROR\n"); return; }
+    if ((status & 0x01) == 0x01) {
+        printf("ATA READ ERROR\n");
+        return;
+    }
     while ((status & 0x08) != 0x08) status = commandPort.Read();
 
     // --- OPTIMIZED READ ---
@@ -99,15 +104,14 @@ void AdvancedTechnologyAttachment::Read28(uint32_t sectorNum, uint8_t* data, int
         // Slow Path
         uint8_t sectorBuffer[512];
         insw(dataPort.getPortNumber(), sectorBuffer, 256);
-        
+
         for (int i = 0; i < count; i++) {
             data[i] = sectorBuffer[i];
         }
     }
 }
 
-void AdvancedTechnologyAttachment::Write28(uint32_t sectorNum, uint8_t* data, uint32_t count)
-{
+void AdvancedTechnologyAttachment::Write28(uint32_t sectorNum, uint8_t* data, uint32_t count) {
     if (sectorNum > 0x0FFFFFFF) return;
     if (count > 512) count = 512;
 
@@ -135,8 +139,10 @@ void AdvancedTechnologyAttachment::Write28(uint32_t sectorNum, uint8_t* data, ui
         // Partial write: Copy to padded buffer first
         uint8_t sectorBuffer[512];
         for (int i = 0; i < 512; i++) {
-            if (i < (int)count) sectorBuffer[i] = data[i];
-            else sectorBuffer[i] = 0;
+            if (i < (int)count)
+                sectorBuffer[i] = data[i];
+            else
+                sectorBuffer[i] = 0;
         }
         outsw(dataPort.getPortNumber(), sectorBuffer, 256);
     }
@@ -144,12 +150,10 @@ void AdvancedTechnologyAttachment::Write28(uint32_t sectorNum, uint8_t* data, ui
     Flush();
 }
 
-void AdvancedTechnologyAttachment::Flush()
-{
-    devicePort.Write( master ? 0xE0 : 0xF0 );
+void AdvancedTechnologyAttachment::Flush() {
+    devicePort.Write(master ? 0xE0 : 0xF0);
     commandPort.Write(0xE7);
     uint8_t status = commandPort.Read();
-    if(status == 0x00) return;
-    while(((status & 0x80) == 0x80) && ((status & 0x01) != 0x01))
-        status = commandPort.Read();
+    if (status == 0x00) return;
+    while (((status & 0x80) == 0x80) && ((status & 0x01) != 0x01)) status = commandPort.Read();
 }
